Reset allowRunningInsecureContent between navigations
Fix #1047

So this re-creates the webview anytime there is a change in the
allowRunningInsecureContent flag.  This is currently equivalent to
everytime you navigate to / from Twitch.  You'll notice that the src
gets set to location anytime this happens now, I think this is more of
the correct behavior because you don't want the old src being loaded you
want the location being loaded.  It also fixes a nasty redirect loop
this way.

Auditors: @diracdeltas